requirements.txt
# streamlit
# opencv-python-headless
# ultralytics
# numpy
# pandas
# opencv-python-headless
streamlit==1.20.0
ultralytics==8.0.20
opencv-python-headless
numpy
pillow
